backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.79.0",
    "fastapi>=0.128.4",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "hypothesis>=6.151.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]


[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
# 1行の最大文字数（Black互換）
line-length = 88

# 対象Pythonバージョン
target-version = "py312"

# 除外するファイル/ディレクトリ
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".hypothesis",
    ".pytest_cache",
]

[tool.ruff.lint]
# 有効にするルール（バランス型）
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import整列)
    "N",    # pep8-naming
    "UP",   # pyupgrade (Python構文の最新化)
    "B",    # flake8-bugbear (バグを見つける)
    "C4",   # flake8-comprehensions (リスト内包表記)
    "SIM",  # flake8-simplify (コード簡略化)
]

# 無視するルール
ignore = [
    "E501",  # 行が長すぎる（formatterに任せる）
]

# ファイルごとに無視するルール
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # テストでassert使用OK
"__init__.py" = ["F401"]    # 未使用import OK

[tool.ruff.format]
# クォートスタイル
quote-style = "double"

# インデント
indent-style = "space"

[tool.ruff.lint.isort]
# 自プロジェクトのパッケージ名
known-first-party = ["app"]
